driver: video: fix b2b switch of power state of video hardware
When the vcodec GDSC is powered on, it moves into HW control. As it
moves into HW control,vcodec is initiated with power down sequence.
Driver then requests for migrating GDSC into sw control, which implies
power up sequence for vcodec. Due to b2b switch of power off and on for
video hardware, it ends up in transient state and hungs eventually.
Writing the register explicitly to avoid power off sequence when HW
control is set.

Change-Id: Ib46c2dc8659b8f68710ac27c7c359255f863bbc6
Signed-off-by: Vikash Garodia <quic_vgarodia@quicinc.com>